Extract `queryStringToParams`
